requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
jinja2==3.1.3
openai==1.30.1
markdown2==2.4.10 